+++ src/app/sample/sample-routing.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { SampleIndexComponent } from './sample-index/sample-index.component';
 import { SampleDatabaseComponent } from './sample-database/sample-database.component';
-import { SampleComponent } from './sample/sample.component';
 import { SampleFormComponent } from './sample-form/sample-form.component';
 import { SampleViewComponent } from './sample-view/sample-view.component';
 import { SgxScaleAuthComponent } from './sgx-scale-auth/sgx-scale-auth.component';